requirements.txt
Bootstrap_Flask==2.3.3
Flask_CKEditor==0.5.1
Flask_Login==0.6.3
Flask-Gravatar==0.5.0
Flask_WTF==1.2.1
WTForms==3.0.1
Werkzeug==3.0.6
Flask==2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv~=1.1.0
humanize~=4.12.2
Flask-Migrate~=4.1.0
pytest~=8.3.5
SQLAlchemy-Utils~=0.41.2
alembic~=1.15.2
